+++ src/MicroOcpp/Model/Transactions/Transaction.h
@@ -5,6 +5,37 @@
 #ifndef TRANSACTION_H
 #define TRANSACTION_H
 
+#include <MicroOcpp/Version.h>
+
+/* General Tx defs */
+#ifdef __cplusplus
+extern "C" {
+#endif //__cplusplus
+
+//TxNotification - event from MO to the main firmware to notify it about transaction state changes
+typedef enum {
+    UNDEFINED,
+
+    //Authorization events
+    TxNotification_Authorized, //success
+    TxNotification_AuthorizationRejected, //IdTag/token not authorized
+    TxNotification_AuthorizationTimeout, //authorization failed - offline
+    TxNotification_ReservationConflict, //connector/evse reserved for other IdTag
+
+    TxNotification_ConnectionTimeout, //user took to long to plug vehicle after the authorization
+    TxNotification_DeAuthorized, //server rejected StartTx/TxEvent
+    TxNotification_RemoteStart, //authorized via RemoteStartTx/RequestStartTx
+    TxNotification_RemoteStop, //stopped via RemoteStopTx/RequestStopTx
+
+    //Tx lifecycle events
+    TxNotification_StartTx, //entered running state (StartTx/TxEvent was initiated)
+    TxNotification_StopTx, //left running state (StopTx/TxEvent was initiated)
+}   TxNotification;
+
+#ifdef __cplusplus
+}
+#endif //__cplusplus
+
 #ifdef __cplusplus
 
 #include <MicroOcpp/Core/Time.h>
@@ -185,8 +216,6 @@ public:
 };
 
 } // namespace MicroOcpp
-
-#include <MicroOcpp/Version.h>
 
 #if MO_ENABLE_V201
 
@@ -422,7 +451,18 @@ extern "C" {
 struct OCPP_Transaction;
 typedef struct OCPP_Transaction OCPP_Transaction;
 
+/*
+ * Compat mode for transactions. This means that all following C-wrapper functions will interprete the handle as v201 transactions
+ */
+#if MO_ENABLE_V201
+void ocpp_tx_compat_setV201(bool isV201); //if set, all OCPP_Transaction* handles are treated as v201 transactions
+#endif
+
 int ocpp_tx_getTransactionId(OCPP_Transaction *tx);
+#if MO_ENABLE_V201
+const char *ocpp_tx_getTransactionIdV201(OCPP_Transaction *tx);
+#endif
+
 bool ocpp_tx_isAuthorized(OCPP_Transaction *tx);
 bool ocpp_tx_isIdTagDeauthorized(OCPP_Transaction *tx);